+++ src/index.js
@@ -9,6 +9,7 @@ import _configure from './api/configure';
 import _call from './api/call';
 import _setGlobalState from './api/set-global-state';
 import _setState from './api/set-state';
+import _set from './api/set';
 import _getState from './api/get-state';
 import _getGlobalState from './api/get-global-state';
 import _getComputed from './api/get-computed';
@@ -40,6 +41,7 @@ export const configure = _configure;
 export const call = _call;
 export const setGlobalState = _setGlobalState;
 export const setState = _setState;
+export const set = _set;
 export const getState = _getState;
 export const getGlobalState = _getGlobalState;
 export const getConnectedState = _getConnectedState;
@@ -77,8 +79,9 @@ const defaultExport = {
   configure,
   call,
   setGlobalState,
-  getGlobalState,
   setState,
+  set,
+  getGlobalState,
   getState,
   getComputed,
   getConnectedState,